backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai>=0.8.0
firebase-admin==6.4.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0 